LPS-194420 refactor TempAttachmentWikiUploadFileEntryHandler.java
moving class to EditPageAttachmentMVCActionCommand.java and deleting handler class
